split files, apply filter on endpoints
